hf_space/requirements.txt
torch
transformers
gradio
transformers 
setfit 
neural_compressor 
optimum[onnxruntime] 
onnxruntime_extensions 
wandb